Change the type of the PasswordInputProperty type's id field from 'Int' to 'Float' since GF stores them in the format '123.456' and they can be cast to a float properly, but not to an int

+++ src/Types/Field/FieldProperty/PasswordInputProperty.php
@@ -25,7 +25,7 @@ class PasswordInputProperty implements Hookable, Type {
             'description' => __('Gravity Forms input property.', 'wp-graphql-gravity-forms'),
             'fields'      => [
                 'id' => [
-                    'type'        => 'Integer',
+                    'type'        => 'Float',
                     'description' => __('The id of the input field.', 'wp-graphql-gravity-forms'),
                 ],
                 'label' => [
